chore(tidy): move join_handler to format/
